Refactor Qcrit handling via dedicated loader
