+++ schemas/qcrit_sram6t.schema.json
